pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.3"
description = "An API serving the Findings Environment for Collected Holdings application"
authors = ["Gregory Wolfe <email@removed.com>"]
readme = "README.md"

[tool.setuptools_scm]
write_to = "api/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"


[tool.poetry.dependencies]
python = "3.11.4"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pytest = "^7.4.1"
httpx = "^0.24.1"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.0.3"
pytz = "^2023.3.post1"
sqlalchemy-utils = "^0.41.1"
pytest-html = "^4.1.1"
python-dotenv = "^1.0.1"
sqlalchemyseed = "^2.0.0"
python3-saml = "1.16.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
sqltap = "^0.3.11"
fastapi-pagination = "0.12.34"
sqlmodel = "0.0.22"
fastapi = "0.115.8"
debugpy = "1.8.12"
setuptools-scm = "8.0.4"
gunicorn = "^23.0.0"
httptools = "^0.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
